# flaskserver/flask_server/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OnTime Attendance Tracking API
  description: "API documentation for OnTime attendance tracking software.This API\
    \ documentation highlights the various requests we would implement using the rest\
    \ API . They are generally shown through put, get and post requests. The attendance\
    \ tracking in general requires database changes on a major amount and hence need\
    \ to be handled well and hence we tested our API documentation using swagger."
  version: 1.0.0
servers:
- url: http://ontime.com/api
paths:
...
  /register/org_info:
    post:
      tags:
      - Organization Registration
      summary: Register a new organization
      operationId: get_org_info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/register_org_info_body'
      responses:
        "200":
          description: The requested action was successful.
        "400":
          description: The passwords do not match.
        "409":
          description: Organisation already registered
      x-openapi-router-controller: swagger_server.controllers.organization_registration_controller
  /register/{organization_id}/mem_info:
    post:
      tags:
      - Organization Registration
      summary: Upload Database of Employees in a file.
      description: "Endpoint for uploading a database file containing details of the\
        \ employee such as its Employee id,Name,Email ID."
      operationId: post_mem_info
      parameters:
      - name: organization_id
        in: path
        description: organization whose members are updated
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/database'
      responses:
        "201":
          description: Database file uploaded successfully
        "404":
          description: Not found
        "403":
          description: Access Denied
      x-openapi-router-controller: swagger_server.controllers.organization_registration_controller
  /register/settings:
    post:
      tags:
      - Organization Registration
      summary: Select Settings for attendance management.
      description: Select the different days which are working days for your organization
        and any ML API you wanted
      operationId: post_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/settings'
        required: true
      responses:
        "200":
          description: Succesful Response
        "404":
          description: No Response found
      x-openapi-router-controller: swagger_server.controllers.organization_registration_controller
...
components:
  schemas:
    database:
      required:
      - Database File
      type: object
      properties:
        Database File:
          maxLength: 10485760
          type: string
          description: The database file containing employee details.
          format: binary
    settings:
      type: object
      properties:
        workingDays:
          type: array
          example:
          - monday
          - tuesday
          - wednesday
          - thursday
          - friday
          items:
            type: string
            enum:
            - monday
            - tuesday
            - wednesday
            - thursday
            - friday
            - saturday
            - sunday
        Face Detection:
          type: boolean
          example: true
        Face Mask Detection:
          type: boolean
          example: false
      example:
        Face Detection: true
        workingDays:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        Face Mask Detection: false
...
    register_org_info_body:
      required:
      - Address
      - Admin Name
      - Confirm Password
      - Email ID
      - Organization ID
      - Password
      - Phone Number
      type: object
      properties:
        Organization ID:
          type: string
          description: The ID of the new organization being registered
        Admin Name:
          type: string
          description: The name of the admin for the new organization being registered
        Email ID:
          type: string
          description: The email address of the admin for the new organization being
            registered
          format: email
        Phone Number:
          type: string
          description: The phone number of the admin for the new organization being
            registered
        Password:
          minLength: 8
          type: string
          description: The password for the new organization being registered
        Confirm Password:
          minLength: 8
          type: string
          description: Confirming password for the new organization being registered
        Address:
          type: string
          description: Address of the Office of the new organization being registered
...